requirements.txt
Django==4.2
pymysql==1.0.2
mysqlclient==2.1.1
apscheduler==3.10.1
django-cors-headers==3.13.0
tencentcloud_sdk_python==3.0.917